Decentralize and improve zsh config
Unlike it was specified in earlier versions, .zprofile is
the true login file - that runs on login. So its jobs is
just start the graphical server.

Also, .zshenv will be always sourced, so it should not
contain commands that produce output or assume the shell
is attached to a TTY.